requirements.txt
# Web Framework
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.0.5

# Data Processing
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
scipy==1.11.1

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.10.0
aiohttp==3.8.5
lxml==4.9.3

# Machine Learning
tensorflow==2.13.0
torch==2.0.1
transformers==4.31.0
joblib==1.3.2

# Database
sqlalchemy==2.0.19
pymysql==1.1.0
redis==4.6.0

# Utilities
python-dotenv==1.0.0
click==8.1.6
tqdm==4.65.0
pyyaml==6.0.1
cryptography==41.0.3

# Monitoring
prometheus-client==0.17.1
structlog==23.1.0

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0